pyproject.toml
[project]
name = "innovatecorp-agent"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.3.0",
    "langchain-aws>=0.2.0",
    "langchain-core>=0.3.0",
    "langsmith>=0.2.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*"]